+++ public/js/basket.js
@@ -427,7 +427,7 @@ export default class extends Controller {
 
     // Updates the currency symbol based on the currency code
     getCurrencySymbol(currencyCode) {
-        if (!currencyCode) { return ''; };
+        if (!currencyCode) { return ''; }
 
         const symbols = {
             "eur": "€",